+++ VirtualConsole/Bus.h
@@ -9,12 +9,12 @@ class Bus
 {
 public:
 
-    void mapDevice(Device* device, uint16_t start, uint16_t end);
+    void mapDevice(Device* device, uint32_t start, uint32_t end);
 
-    uint8_t read(uint16_t address);
-    void write(uint16_t address, uint8_t value);
+    uint8_t read(uint32_t address);
+    void write(uint32_t address, uint8_t value);
 
-    bool isMapped(uint16_t address);
+    bool isMapped(uint32_t address);
 
     void tick();
 
@@ -25,11 +25,11 @@ private:
     struct MappedDevice
     {
         Device* device;
-        uint16_t start;
-        uint16_t end;
+        uint32_t start;
+        uint32_t end;
     };
 
     std::vector<MappedDevice> devices;
 
-    MappedDevice* findDevice(uint16_t address);
+    MappedDevice* findDevice(uint32_t address);
 };
