+++ VirtualConsole/Keyboard.h
@@ -8,8 +8,8 @@ public:
 
     Keyboard();
 
-    uint8_t read(uint16_t address) override;
-    void write(uint16_t address, uint8_t value) override;
+    uint8_t read(uint32_t address) override;
+    void write(uint32_t address, uint8_t value) override;
 
     void tick() override;
