+++ VirtualConsole/TextVRAM.h
@@ -12,8 +12,8 @@ public:
 
     TextVRAM();
 
-    uint8_t read(uint16_t address) override;
-    void write(uint16_t address, uint8_t value) override;
+    uint8_t read(uint32_t address) override;
+    void write(uint32_t address, uint8_t value) override;
 
     void render() const;
